Extract out SiteNestedContainer to its own file
